changed to double

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -16,7 +16,9 @@ public class MyMathCalc {
     public int multiply(int a,int b){
         return 0;
     }
-    public float divide(int a,int b){
+
+    public double divide(int a,int b){
+  
 
         return 0;
     }
